Added methods and obects for client status
Library needs some way to store what state the connected GUI
is in. Clients are going to be used as data holders for the GUICommandDisplays

New connecting clients are now added to client list.
Disconnecting clients are removed from the list
Clients that are connected can have thier status string updated by the library